+++ lib/src/analyzers/lint_analyzer/reporters/reporters_list/github/lint_github_reporter.dart
@@ -4,6 +4,7 @@ import '../../../../../reporters/models/github_reporter.dart';
 import '../../../models/lint_file_report.dart';
 import '../../../models/severity.dart';
 import '../../../models/summary_lint_report_record.dart';
+import '../../lint_report_params.dart';
 
 const _deprecationMessage =
     'DEPRECATED! This reporter is deprecated and will be removed in 5.0.0. You can migrate on our GitHub Action.';
@@ -15,14 +16,14 @@ const _deprecationMessage =
 /// **Note:** this reporter is deprecated and should not be used.
 /// Consider using Dart Code Metrics GitHub Action instead.
 class LintGitHubReporter extends GitHubReporter<LintFileReport,
-    SummaryLintReportRecord<Object>, void> {
+    SummaryLintReportRecord<Object>, LintReportParams> {
   const LintGitHubReporter(IOSink output) : super(output);
 
   @override
   Future<void> report(
     Iterable<LintFileReport> records, {
     Iterable<SummaryLintReportRecord<Object>> summary = const [],
-    void additionalParams,
+    LintReportParams? additionalParams,
   }) async {
     if (records.isEmpty) {
       return;