+++ lib/src/analyzers/unused_code_analyzer/reporters/reporter_factory.dart
@@ -4,14 +4,18 @@ import '../../../../unused_code_analyzer.dart';
 import '../../../reporters/models/console_reporter.dart';
 import '../../../reporters/models/json_reporter.dart';
 import '../../../reporters/models/reporter.dart';
+import 'unused_code_report_params.dart';
 
-final _implementedReports = <String,
-    Reporter<UnusedCodeFileReport, void, void> Function(IOSink output)>{
+final _implementedReports = <
+    String,
+    Reporter<UnusedCodeFileReport, void, UnusedCodeReportParams> Function(
+  IOSink output,
+)>{
   ConsoleReporter.id: (output) => UnusedCodeConsoleReporter(output),
   JsonReporter.id: (output) => UnusedCodeJsonReporter(output),
 };
 
-Reporter<UnusedCodeFileReport, void, void>? reporter({
+Reporter<UnusedCodeFileReport, void, UnusedCodeReportParams>? reporter({
   required String name,
   required IOSink output,
 }) {