+++ lib/src/analyzers/unused_code_analyzer/reporters/reporters_list/json/unused_code_json_reporter.dart
@@ -4,19 +4,20 @@ import 'dart:io';
 import '../../../../../reporters/models/json_reporter.dart';
 import '../../../models/unused_code_file_report.dart';
 import '../../../models/unused_code_issue.dart';
+import '../../unused_code_report_params.dart';
 
 /// Unused code JSON reporter.
 ///
 /// Use it to create reports in JSON format.
 class UnusedCodeJsonReporter
-    extends JsonReporter<UnusedCodeFileReport, void, void> {
+    extends JsonReporter<UnusedCodeFileReport, void, UnusedCodeReportParams> {
   const UnusedCodeJsonReporter(IOSink output) : super(output, 2);
 
   @override
   Future<void> report(
     Iterable<UnusedCodeFileReport> records, {
     Iterable<void> summary = const [],
-    void additionalParams,
+    UnusedCodeReportParams? additionalParams,
   }) async {
     if (records.isEmpty) {
       return;