+++ lib/src/analyzers/unused_code_analyzer/reporters/unused_code_report_params.dart
@@ -1,0 +1,6 @@
+/// Represents additional unused code reporter params
+class UnusedCodeReportParams {
+  final bool congratulate;
+
+  const UnusedCodeReportParams({required this.congratulate});
+}