+++ lib/src/analyzers/unused_code_analyzer/unused_code_analyzer.dart
@@ -20,6 +20,7 @@ import 'models/unused_code_file_report.dart';
 import 'models/unused_code_issue.dart';
 import 'public_code_visitor.dart';
 import 'reporters/reporter_factory.dart';
+import 'reporters/unused_code_report_params.dart';
 import 'unused_code_analysis_config.dart';
 import 'unused_code_config.dart';
 import 'used_code_visitor.dart';
@@ -30,7 +31,7 @@ class UnusedCodeAnalyzer {
 
   /// Returns a reporter for the given [name]. Use the reporter
   /// to convert analysis reports to console, JSON or other supported format.
-  Reporter<UnusedCodeFileReport, void, void>? getReporter({
+  Reporter<UnusedCodeFileReport, void, UnusedCodeReportParams>? getReporter({
     required String name,
     required IOSink output,
   }) =>
@@ -88,7 +89,9 @@ class UnusedCodeAnalyzer {
       }
     }
 
-    codeUsages.exports.forEach(publicCode.remove);
+    if (!config.isMonorepo) {
+      codeUsages.exports.forEach(publicCode.remove);
+    }
 
     return _getReports(codeUsages, publicCode, rootFolder);
   }
@@ -114,10 +117,12 @@ class UnusedCodeAnalyzer {
     String rootFolder,
     Iterable<Glob> excludes,
   ) {
-    final contextFolders = folders
-        .where((path) => normalize(join(rootFolder, path))
-            .startsWith(context.contextRoot.root.path))
-        .toList();
+    final contextFolders = folders.where((path) {
+      final newPath = normalize(join(rootFolder, path));
+
+      return newPath == context.contextRoot.root.path ||
+          context.contextRoot.root.path.startsWith('$newPath/');
+    }).toList();
 
     return extractDartFilesFromFolders(contextFolders, rootFolder, excludes);
   }