+++ test/src/analyzers/lint_analyzer/reporters/reporters_list/console/lint_console_reporter_helper_test.dart
@@ -68,6 +68,11 @@ void main() {
 
     test('getMetricMessage returns formatted message', () {
       expect(
+        helper.getMetricMessage(MetricValueLevel.alarm, 'Class.method', []),
+        isEmpty,
+      );
+
+      expect(
         helper.getMetricMessage(
           MetricValueLevel.alarm,
           'Class.method',