+++ test/src/analyzers/unused_files_analyzer/reporters/reporters_list/console/unused_files_console_reporter_test.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:ansicolor/ansicolor.dart';
 import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/models/unused_files_file_report.dart';
 import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/reporters/reporters_list/console/unused_files_console_reporter.dart';
+import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/reporters/unused_files_report_params.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
@@ -12,42 +13,91 @@ void main() {
   group('UnusedFilesConsoleReporter reports', () {
     const fullPath = '/home/developer/work/project/example.dart';
 
-    late IOSinkMock _output; // ignore: close_sinks
-    late UnusedFilesConsoleReporter _reporter;
+    late IOSinkMock output; // ignore: close_sinks
+
+    late UnusedFilesConsoleReporter reporter;
 
     setUp(() {
-      _output = IOSinkMock();
+      output = IOSinkMock();
 
       ansiColorDisabled = false;
-      _reporter = UnusedFilesConsoleReporter(_output);
+
+      reporter = UnusedFilesConsoleReporter(output);
     });
 
-    test('no unused files', () async {
-      await _reporter.report([]);
+    group('empty report', () {
+      test('with congratulate param', () async {
+        const congratulate = UnusedFilesReportParams(
+          congratulate: true,
+          deleteUnusedFiles: false,
+        );
 
-      final report =
-          verify(() => _output.writeln(captureAny())).captured.cast<String>();
+        await reporter.report([], additionalParams: congratulate);
 
-      expect(report, equals(['\x1B[38;5;20m✔\x1B[0m no unused files found!']));
+        final captured =
+            verify(() => output.writeln(captureAny())).captured.cast<String>();
+        expect(
+          captured,
+          equals(['\x1B[38;5;20m✔\x1B[0m no unused files found!']),
+        );
+      });
+
+      test('without congratulate param', () async {
+        const noCongratulate = UnusedFilesReportParams(
+          congratulate: false,
+          deleteUnusedFiles: false,
+        );
+
+        await reporter.report([], additionalParams: noCongratulate);
+
+        verifyNever(() => output.writeln(any()));
+      });
     });
 
-    test('unused files', () async {
-      const record =
-          UnusedFilesFileReport(path: fullPath, relativePath: 'example.dart');
+    group('unused files', () {
+      test('without deleting flag', () async {
+        const record =
+            UnusedFilesFileReport(path: fullPath, relativePath: 'example.dart');
 
-      await _reporter.report([record]);
+        await reporter.report([record]);
 
-      final report =
-          verify(() => _output.writeln(captureAny())).captured.cast<String>();
+        final captured =
+            verify(() => output.writeln(captureAny())).captured.cast<String>();
 
-      expect(
-        report,
-        equals([
-          '\x1B[38;5;180m⚠\x1B[0m unused file: example.dart',
-          '',
-          '\x1B[38;5;167m✖\x1B[0m total unused files - \x1B[38;5;167m1\x1B[0m',
-        ]),
-      );
+        expect(
+          captured,
+          equals([
+            '\x1B[38;5;180m⚠\x1B[0m unused file: example.dart',
+            '',
+            '\x1B[38;5;167m✖\x1B[0m total unused files - \x1B[38;5;167m1\x1B[0m',
+          ]),
+        );
+      });
+
+      test('with deleting flag', () async {
+        const record =
+            UnusedFilesFileReport(path: fullPath, relativePath: 'example.dart');
+
+        await reporter.report(
+          [record],
+          additionalParams: const UnusedFilesReportParams(
+            congratulate: true,
+            deleteUnusedFiles: true,
+          ),
+        );
+
+        final captured =
+            verify(() => output.writeln(captureAny())).captured.cast<String>();
+
+        expect(
+          captured,
+          equals([
+            '\x1B[38;5;180m⚠\x1B[0m unused file: example.dart',
+            '',
+            '\x1B[38;5;20m✔\x1B[0m \x1B[38;5;167m1\x1B[0m files were successfully deleted',
+          ]),
+        );
+      });
     });
   });
 }